agents/requirements.txt
google-api-python-client
python-dotenv
python-pptx
requests
beautifulsoup4
